Add new function for graceful degredation of a preview panel.
